further beautified UI and added lobby
There is still code to be added that removes a player from a game if they 'join game' and then 'leave lobby'. Also, the timer starts once the lobby is made and not once 'start game' is pressed. ALSO, one player clicking 'start game' should cause all UI's to go to the game but currently does not.